(De)serializing any type from a (de)serializable type
I found myself relying on this trick again today but couldn't find it documented anywhere.

It kind of follows from the design of serde, but I suspect there are a lot of users that are not aware of this and might do something more complicated, e.g. implement a Visitor like described in the docs.